backend/adk-bidi/pyproject.toml
[project]
name = "athena-backend"
version = "0.1.0"
description = "Athena ADK bidi streaming backend — voice + screen"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }

dependencies = [
    "google-adk>=1.20.0",
    "langsmith[google-adk]>=0.7.17",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.32.0",
    "greenlet>=3.0.0",
    "pyyaml>=6.0.0",
    "aiosqlite>=0.20.0",
    "numpy>=1.26.0",
    "nb-cli-tool>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
